Transfers is now supported for all tokens with a .0000 precision Fixed a critical issue where sequence responses were all being parsed as Int, rather than Long

+++ android/app/src/main/java/com/memtrip/eosreach/api/actions/AccountActionList.kt
@@ -20,5 +20,5 @@ import com.memtrip.eosreach.api.actions.model.AccountAction
 
 data class AccountActionList(
     val actions: List<AccountAction>,
-    val noResultsNext: Int = -1
+    val noResultsNext: Long = -1
 )
